# src/configs/swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {
      "CreateTest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["email", "name"],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "dej": {
        "description": "check profile"
      }
    }
  },
  "info": {
    "title": "Express Template API",
    "version": "1.0.0",
    "description": "API documentation",
    "license": {
      "name": "ISC"
    },
    "contact": {}
  },
  "paths": {
    "/test": {
      "post": {
        "operationId": "Create",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateTest"
                    }
                  },
                  "required": ["message", "data"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Test"],
        "security": [
          {
            "dej": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/CreateTest"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["message", "data"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Test"],
        "security": [
          {
            "dej": []
          }
        ],
        "parameters": []
      }
    },
    "/test/{id}": {
      "get": {
        "operationId": "Read",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateTest"
                    }
                  },
                  "required": ["message", "data"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Test"],
        "security": [
          {
            "dej": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      },
      "delete": {
        "operationId": "Remove",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Test"],
        "security": [
          {
            "dej": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      },
      "put": {
        "operationId": "Update",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateTest"
                    }
                  },
                  "required": ["message", "data"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Test"],
        "security": [
          {
            "dej": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTest"
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://192.168.219.104:3000/"
    }
  ]
}